feat: add support for beta channel wine releases
fix: detect #435 and suggest changing wine channel.
Should we force this?

chore: updated "Get Support" text to add reminder about checking pinned issues before reporting

refactor: store wine_binary as a sigil rather than a path for part of the installation so we can detect the difference between needing to download something and a file on the system (as well as figure out if the user just wants the recommended and doesn't really care about the details)

refactor: change the API endpoint to get all the releases so that we can get the prerelease ones as well

feat(TUI): add button to toggle wine release channel.
Should this be on other UIs as well, especially if users are forced down this path to avoid issue #435? Something doesn't smell right UX wise.