feat: support wine pre releases (#437)
* fix: detect chromium and avoid using it

There is presently a wine bug that makes chromium as a browser not work. Detect and remediate.

Also run update-desktop-database after install. Needed for snaps to pickup the new mime type.

* feat: add support for beta channel wine releases

fix: detect #435 and suggest changing wine channel.
Should we force this?

chore: updated "Get Support" text to add reminder about checking pinned issues before reporting

refactor: store wine_binary as a sigil rather than a path for part of the installation so we can detect the difference between needing to download something and a file on the system (as well as figure out if the user just wants the recommended and doesn't really care about the details)

refactor: change the API endpoint to get all the releases so that we can get the prerelease ones as well

feat(TUI): add button to toggle wine release channel.
Should this be on other UIs as well, especially if users are forced down this path to avoid issue #435? Something doesn't smell right UX wise.

* ci: fix tests

* fix: change switching to beta wine to be a one-time operation during install

* fix: integration bugs

remove non-linux browsers
fix applications path
handle case where the wine binary is stored as the path rather than the sigil

Tested:
- debian 12 snap firefox
- install correctly switch to -devel
- (install failed due to a fuse issue, unrelated)

- debian 12 flatpak firefox
- install correctly switch to -devel
- (installed failed due to running out of space in VM)

- debian 13 Dooble correctly switched to -devel
- aborted install once text was obsered

- debian 13 Opera correctly switched to -devel
- aborted install once text was obsered

- debian 13 Chromium correctly switched to -devel
- aborted install once text was obsered

- debian 13 Qutebrowser correctly switched to -devel
- aborted install once text was obsered

- debian 13 Vivaldi correctly switched to -devel
- aborted install once text was obsered

- debian 13 Brave correctly switched to -devel
- aborted install once text was obsered

- debian 13 Falkon correctly switched to -devel
- aborted install once text was obsered

- debian 13 Google Chrome correctly switched to -devel
- aborted install once text was obsered

- debian 12 Chromium correctly switched to -devel
- install success
- logged in
- got to the "Download your Logos Library page"

* fix: issue if install was run again it switched from beta to stable

Reproduction:
- install once needing the -devel
- run the install again
- notice how it switched back to -staging

Tested:
- Debian chromium install once (-devel)
- re-ran install (observed -devel)